commit clase

+++ app/src/main/java/com/iessanalbertomagno/dam2/maquinadecambio/data/Dinero.kt
@@ -6,6 +6,6 @@ data class Dinero(
     var idImagen : Int,
     var descripcion: String,
     var tamanno: Dp,
-    var valor: Int,
+    var valor: Double,
     var cantidad: Int
     )
